+++ teca/src/app/register/components/avatar-select/professor-avatar/avatar/avatar.component.scss
