--- teca/src/app/register/components/avatar-select/professor-avatar/man-avatar/man-avatar.component.html
@@ -1,29 +1,0 @@
-<ion-grid>
-  <ion-row>
-    <ion-col>
-      <ion-segment (ionChange)="selectAvatar($event)">
-        <ion-segment-button value="assets/img/avatar/male/professor/professor1.png">
-          <ion-img src="assets/img/avatar/male/professor/professor1.png"></ion-img>
-      </ion-segment-button>
-
-      <ion-segment-button value="assets/img/avatar/male/professor/professor2.png">
-        <ion-img src="assets/img/avatar/male/professor/professor2.png"></ion-img>
-      </ion-segment-button>
-
-      <ion-segment-button value="assets/img/avatar/male/professor/professor3.png">
-        <ion-img src="assets/img/avatar/male/professor/professor3.png"></ion-img>
-      </ion-segment-button>  
-
-      <ion-segment-button value="assets/img/avatar/male/professor/professor4.png">
-        <ion-img src="assets/img/avatar/male/professor/professor4.png"></ion-img>
-      </ion-segment-button>        
-    </ion-segment>    
-    </ion-col>
-  </ion-row> 
-
-  <ion-row>
-    <ion-col class="btnEmpecemos">
-      <ion-button shape="round" (click)="create()">{{comando}}</ion-button>
-    </ion-col>
-  </ion-row>    
-</ion-grid>
