--- teca/src/app/register/components/avatar-select/professor-avatar/man-avatar/man-avatar.component.scss
@@ -1,21 +1,0 @@
-.avatar{  
-    height: 150px;
-    width: 150px;
-    
-}
-.btnAvatar{
-    //position: absolute;
-    width: 139px;
-    height: 129px;
-    //left: 576px;
-    top: 20px;    
-    text-align: center;
-}
-.btnEmpecemos{
-    //position: absolute;
-    width: 215px;
-    height: 47px;
-    //left: 780px;
-    top: 80px;    
-    text-align: center;
-}
