+++ teca/src/app/register/components/avatar-select/professor-avatar/professor-avatar.component.html
@@ -7,8 +7,7 @@
 
   <ion-row >
     <ion-col>
-      <app-woman-avatar comando="Comencemos" (newWomanEvent)="create($event)" *ngIf="this.visibleFemenino"></app-woman-avatar>
-      <app-man-avatar comando="Comencemos" (newManEvent)="create($event)" *ngIf="this.visibleMasculino"></app-man-avatar>
+      <app-woman-avatar comando="Comencemos" (newWomanEvent)="create($event)" ></app-woman-avatar>
     </ion-col>
   </ion-row>
