+++ teca/src/app/register/register.module.ts
@@ -8,10 +8,9 @@ import { RegisterFormComponent } from './components/register-form/register-form.
 import { FormsModule } from '@angular/forms';
 import { AvatarSelectComponent } from './components/avatar-select/avatar-select.component';
 import { GirlAvatarComponent } from './components/avatar-select/student-avatar/avatar/avatar.component';
-import { ManAvatarComponent } from './components/avatar-select/professor-avatar/man-avatar/man-avatar.component';
 import { ProfessorAvatarComponent } from './components/avatar-select/professor-avatar/professor-avatar.component';
 import { StudentAvatarComponent } from './components/avatar-select/student-avatar/student-avatar.component';
-import { WomanAvatarComponent } from './components/avatar-select/professor-avatar/woman-avatar/woman-avatar.component';
+import { WomanAvatarComponent } from './components/avatar-select/professor-avatar/avatar/avatar.component';
 
 
 
@@ -23,7 +22,7 @@ import { WomanAvatarComponent } from './components/avatar-select/professor-avata
     , AvatarSelectComponent
 
     , GirlAvatarComponent
-    , ManAvatarComponent
+
     , ProfessorAvatarComponent
     , StudentAvatarComponent
     , WomanAvatarComponent],
@@ -34,7 +33,7 @@ import { WomanAvatarComponent } from './components/avatar-select/professor-avata
     , AvatarSelectComponent
 
     , GirlAvatarComponent
-    , ManAvatarComponent
+
     , ProfessorAvatarComponent
     , StudentAvatarComponent
     , WomanAvatarComponent],
